Rename Category Service files and update imports
