+++ src/app/components/app-main/app-main.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { CategoryService } from 'src/app/services/data-service.service';
+import { CategoryService } from 'src/app/services/category-service.service';
 
 @Component({
   selector: 'app-app-main',
